Buffered io
```
name                          old time/op    new time/op    delta
Decode/bep/imagemeta/exif-10     212µs ± 4%      24µs ± 3%  -88.46%  (p=0.029 n=4+4)
Decode/bep/imagemeta/all-10      269µs ± 2%      32µs ± 2%  -88.09%  (p=0.029 n=4+4)

name                          old alloc/op   new alloc/op   delta
Decode/bep/imagemeta/exif-10    7.57kB ± 0%    7.77kB ± 0%   +2.67%  (p=0.029 n=4+4)
Decode/bep/imagemeta/all-10     9.08kB ± 0%    9.47kB ± 0%   +4.29%  (p=0.029 n=4+4)

name                          old allocs/op  new allocs/op  delta
Decode/bep/imagemeta/exif-10       832 ± 0%       837 ± 0%   +0.60%  (p=0.029 n=4+4)
Decode/bep/imagemeta/all-10      1.08k ± 0%     1.09k ± 0%   +0.93%  (p=0.029 n=4+4)
```
